src/perlib/requirements.txt
Flask==2.2.2
joblib==1.1.1
keras==2.9.0
keras_tcn==3.1.1
lightgbm==3.2.1
loguru==0.6.0
matplotlib==3.6.1
numpy==1.23.4
pandas==1.5.0
python_dateutil==2.8.2
scikit_learn==1.2.0
scipy==1.9.3
statsmodels==0.13.2
tensorflow==2.11.0
tensorflow_macos==2.10.0
tqdm==4.64.0
xgboost==1.6.2
joblib==1.1.1
